linted most files. Switched to .jsx, created user container
