+++ src/client/actions/index.js
@@ -2,11 +2,11 @@
 
 export const CREATE_USER = 'CREATE_USER';
 
-export function createUser(name) {
+export const createUser = (name) => {
   // some action here that will register the user input in the database
-  console.log('inside the first action');
+
   return {
     type: CREATE_USER,
-    payload: 'a string for now'
-  }
+    payload: 'a string for now',
+  };
 };
