--- src/client/components/landingPage.js
@@ -1,14 +1,0 @@
-import React, {Component} from 'react';
-import User from '../containers/userName';
-
-class Landing extends Component {
-  render() {
-    return (
-      <div>
-        Our baby is born!
-      </div>
-    );
-  }
-}
-
-export default Landing;
