+++ src/client/components/landingPage.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import User from '../containers/userName';
+
+const Landing = () => {
+  return (
+    <div>
+      Our baby is born!
+
+      <User />
+    </div>
+  );
+};
+
+export default Landing;
