--- src/client/containers/userName.js
@@ -1,16 +1,0 @@
-import React, {Component} from 'react';
-import {connect} from 'react-redux';
-import {bindActionCreators} from 'redux';
-import {createUser} from '../actions/index';
-
-class User extends Component {
-  render() {
-    return (
-      <div>
-        another baby
-      </div>
-    )
-  }
-}
-
-export default User;
