+++ src/client/reducers/user.js
@@ -1,13 +1,14 @@
-import {CREATE_USER} from '../actions/index';
-const INITIAL_STATE = {messages: [], user: null};
+import { CREATE_USER } from '../actions/index';
+
+const INITIAL_STATE = { messages: [], user: null };
 
 export default (state = INITIAL_STATE, action) => {
-  switch(action.type) {
+  switch (action.type) {
     case CREATE_USER: {
-      return {...state, user: action.payload.data};
+      return { ...state, user: action.payload.data };
     }
     default: {
       return state;
     }
   }
-}
+};
