--- src/srcTemplate/index.html
@@ -1,18 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link href="styles/style.css" rel="stylesheet">
-  <!--<link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">-->
-  <title>clack!!</title>
-</head>
-
-<body>
-  <div id="container"></div>
-</body>
-<script src="/vendor.js"></script>
-<script src="/app.js"></script>
-
-</html>
